+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version Information
 ## version 2.0.6
+  - add r_runner to shell_methods
 ## version 2.0.5 6193f4c 512022-935
   - Add Float.eround
   - Add Array.eround